--- src/containers/index.js
@@ -1,11 +1,0 @@
-import ToolListContainer from './ToolListContainer';
-import AddToolContainer from './AddToolContainer'
-import UpdateToolContainer from './UpdateToolContainer'
-import AddUpdateFormContainer from './AddUpdateFormContainer'
-import PagingListContainer from './PagingListContainer'
-
-export { ToolListContainer }
-export { AddToolContainer }
-export { UpdateToolContainer }
-export { AddUpdateFormContainer }
-export { PagingListContainer }
