+++ src/core/store.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux'
-import reducer from '../reducer'
+import reducer from './root-reducer'
 
 const store = createStore(reducer)
