+++ src/features/shared/constants/Locations.js
