+++ src/features/shared/index.js
@@ -1,0 +1,6 @@
+import { Actions, Headers, Locations, Types } from './constants'
+
+export {Actions};
+export {Headers};
+export {Locations};
+export {Types};
