+++ src/features/tool-search/shared/add-edit-form/index.js
@@ -1,0 +1,9 @@
+import { changeName, changeType, changeLocation } from './actions'
+import AddUpdateFormComponent from './component'
+import AddUpdateFormContainer from './container'
+
+export { changeName }
+export { changeType }
+export { changeLocation }
+export { AddUpdateFormComponent }
+export { AddUpdateFormContainer }
