+++ src/features/tool-search/tool-add/actions.js
@@ -1,0 +1,12 @@
+import { Actions } from '../../shared/constants'
+
+export const addTool  = ({name, typeTool, location}) => {
+    return {
+        type: Actions.ADD_TOOL,
+        payload: {
+            name,
+            typeTool,
+            location
+        }
+    }
+}
