+++ src/features/tool-search/tool-add/index.js
@@ -1,0 +1,5 @@
+import AddToolContainer from './container'
+import { addTool } from './actions'
+
+export { AddToolContainer }
+export { addTool }
