+++ src/features/tool-search/tool-edit/actions.js
@@ -1,0 +1,13 @@
+import { Actions } from '../../shared/constants'
+
+export const updateTool = ({id, name, typeTool, location}) => {
+    return {
+        type: Actions.UPDATE_TOOL,
+        payload: {
+            id,
+            name,
+            typeTool,
+            location
+        }
+    }
+}
