+++ src/features/tool-search/tool-edit/index.js
@@ -1,0 +1,5 @@
+import UpdateToolContainer from './container'
+import { updateTool } from './actions'
+
+export { UpdateToolContainer }
+export { updateTool }
