+++ src/features/tool-search/tool-list/actions.js
@@ -1,0 +1,8 @@
+import { Actions } from '../../shared/constants'
+
+export const changeCurrentPage = (newPage) => {
+    return {
+        type: Actions.CHANGE_CURRENT_PAGE,
+        payload: newPage
+    }
+}
