+++ src/features/tool-search/tool-list/components/index.js
@@ -1,11 +1,9 @@
 import ToolComponent from './ToolComponent'
 import ToolTableComponent from './ToolTableComponent'
 import ToolListComponent from './ToolListComponent'
-import AddUpdateFormComponent from './AddUpdateFormComponent'
 import PagingListComponent from './PagingListComponent'
 
 export { ToolComponent }
 export { ToolTableComponent }
 export { ToolListComponent }
-export { AddUpdateFormComponent }
 export { PagingListComponent }
