+++ src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {render} from 'react-dom';
-import Root from './Root';
+import Root from './core/root-component';
 
 render(<Root />, document.getElementById('container'));
-
