--- src/reducer/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux'
-import toolsReducer from './toolsReducer'
-import formReducer from './formReducer'
-
-export default combineReducers ({
-    toolsReducer,
-    formReducer
-})
